requirements.txt
openai==1.60.2
langchain-community==0.3.16
langchain-core==0.3.32
langchain_huggingface==0.1.2
huggingface_hub==0.28.0
faiss-cpu==1.9.0.post1
langchain_groq==0.2.3
unstructured==0.16.16
libmagic==1.0
python-magic==0.4.27
streamlit